Implement focus_move() and focus_swap()

+++ lib/X11/korgwm/API.pm
@@ -19,6 +19,7 @@ my $server;
 
 # Establish API server
 sub init {
+    return unless defined $cfg->{api_host} and defined $cfg->{api_port};
     $server = tcp_server $cfg->{api_host}, $cfg->{api_port}, sub($fh, @) {
         my $hdl;
         my $close = sub { $hdl->destroy(); undef $fh; delete $clients{$hdl}; };
